# Conflicts:
#	src/test/java/org/prebid/server/ApplicationTest.java
#	src/test/resources/org/prebid/server/ApplicationTest/openrtb2/adform/test-auction-adform-response.json
#	src/test/resources/org/prebid/server/ApplicationTest/openrtb2/adtelligent/test-auction-adtelligent-response.json
#	src/test/resources/org/prebid/server/ApplicationTest/openrtb2/openx/test-auction-openx-response.json
#	src/test/resources/org/prebid/server/ApplicationTest/openrtb2/pubmatic/test-auction-pubmatic-response.json
#	src/test/resources/org/prebid/server/ApplicationTest/openrtb2/pubmatic/test-cache-pubmatic-request.json
#	src/test/resources/org/prebid/server/ApplicationTest/openrtb2/rubicon_appnexus/test-auction-rubicon-appnexus-response.json
#	src/test/resources/org/prebid/server/ApplicationTest/openrtb2/somoaudience/test-auction-somoaudience-response.json

+++ src/main/java/com/iab/openrtb/request/Imp.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Value;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -92,7 +93,7 @@ public class Imp {
     String tagid;
 
     /** Minimum bid for this impression expressed in CPM. */
-    Float bidfloor;
+    BigDecimal bidfloor;
 
     /**
      * Currency specified using ISO-4217 alpha codes. This may be different from
